Fixed defocus issue
The error occurs when pressing keyboard enter as opposed to software keyboard enter or the submit button. I'm unsure of how this occurs but since the issue isn't going experienced on an actual device, I'll leave it as is.

The weird defocus on back is removed, and a unit test corrected.